Changed file organization for ResidenceAddHeader component.
