+++ Mobile/src/Component/ResidenceAddHeader/index.js
@@ -3,40 +3,7 @@ import React from 'react';
 import {View, Text, StyleSheet, useWindowDimensions} from 'react-native';
 import {Divider} from 'react-native-elements';
 
-const styles = StyleSheet.create({
-  header: {
-    backgroundColor: '#7E57C2',
-
-    justifyContent: 'center',
-    alignItems: 'center',
-
-    padding: 20,
-  },
-  headerTitle: {
-    fontSize: 25,
-    color: '#F9F8FD',
-
-    marginBottom: 15,
-
-    fontWeight: 'bold',
-    fontFamily: 'Roboto',
-  },
-  headerText: {
-    fontSize: 18,
-    color: '#F9F8FD',
-
-    marginBottom: 15,
-
-    fontWeight: 'bold',
-    fontFamily: 'Roboto',
-  },
-  headerDivider: {
-    backgroundColor: '#F9F8FD',
-    height: 3,
-
-    alignSelf: 'center',
-  },
-});
+import styles from './styles';
 export default function ResidenceAddHeader({title, subtitle}) {
   const width = useWindowDimensions().width;
