+++ Mobile/src/Component/ResidenceAddHeader/styles.js
@@ -1,0 +1,35 @@
+const styles = StyleSheet.create({
+  header: {
+    backgroundColor: '#7E57C2',
+
+    justifyContent: 'center',
+    alignItems: 'center',
+
+    padding: 20,
+  },
+  headerTitle: {
+    fontSize: 25,
+    color: '#F9F8FD',
+
+    marginBottom: 15,
+
+    fontWeight: 'bold',
+    fontFamily: 'Roboto',
+  },
+  headerText: {
+    fontSize: 18,
+    color: '#F9F8FD',
+
+    marginBottom: 15,
+
+    fontWeight: 'bold',
+    fontFamily: 'Roboto',
+  },
+  headerDivider: {
+    backgroundColor: '#F9F8FD',
+    height: 3,
+
+    alignSelf: 'center',
+  },
+});
+export default styles;
